Fix UI Proxy issue in Firefox Config and Upgrade to Selenium 4.8.1 (#27)
* Upgrade of dependencies.
Removal of deprecated Phantomjs.
Removal of deprecated code.
Addition of "SuppressWarnings" tags where needed. To de resolved at a later stage.

* Adding IE capabilities for RemoteWebDriver.

* Upgrade of function call.

* Addition of field in the UI to set the path to the Firefox Gecko driver.

* Wiki update.

* Addition of input field for msegde.exe

* Clean-ups and finalisations.

* Addition of IE11 in MsEdge for Remote execution.

* Integrating changes from Shinu on Aug 26, 2019.

* Merging of thefuckingcode changes from Apr 17, 2021.

* Setting of default value of path to msedge.exe

* Update of Chrome tests to JDK 17.

* Update of Firefox tests to JDK 17.

* Upgrade of HTMLUntit tests to JDK 17.

* Update of InternetExplorer tests to JDK 17.

* Update of RemoteDriver tests to JDK 17.

* Update of WebDriver tests to JDK 17.

* Formatting fixes.

* Formatting fixes 2.

* Formatting fixes #3.

* Addition of Geckodriver v0.31 for Firefox tests.

* Usage of GeckoDriver default service for unit tests instead of real Geckodriver.exe.

* Update to Selenium 4.5

* Ui fix.

* Addition of geckodriver to travis CI for the firefox tests

* Matching Continous Integration test conditions.

* Setting IE's Protected Mode to "off" to allow IE to start in all environments.

* Removaĺ of path to Edge as from IEDriver v4.5.0, the driver will
 automatically locate Edge on the system.

* Ui Refactor 1

* Update to Seleniuṁ 4.6.0

* UI Refactor

* Adding of new wiki page for direct browseŕ testing.

* Updating wiki pages for new UI.

* Tests update 1.

* Update of Chrome and IE tests.

* Pom update.

* Imporţclean-ups.

* Update of Readme.md.

* Fix of broken wiki links.

* Fix of broken link #2.

* Take #3

* Code format

* Code format #2

* Syntax highlight.

* Initial Browser URL now in UI and not hard coded.

* Test fixes.

* Upgrade to Selenium 4.7.2 and addition of driver for Edge.

* Update of tests.

* Removing "ignoring zoom levels" in IE as it is no longer required from Seleniuṁ 4.7.0.

* Upgrade to Selenium 4.8.0

* Proxy UI fix in Firefox && upgrade to Selenium 4.8.1.
