+++ pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>kg.apc</groupId>
     <artifactId>jmeter-plugins-webdriver</artifactId>
-    <version>4.8.0</version>
+    <version>4.8.1</version>
     <packaging>jar</packaging>
 
     <name>WebDriver/Selenium</name>
@@ -34,7 +34,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
-        <selenium.version>4.8.0</selenium.version>
+        <selenium.version>4.8.1</selenium.version>
     </properties>
 
     <build>
@@ -136,7 +136,7 @@
         <dependency>
             <groupId>org.seleniumhq.selenium</groupId>
             <artifactId>htmlunit-driver</artifactId>
-            <version>4.7.0</version>
+            <version>4.8.0</version>
         </dependency>
 
         <!-- begin need this libraries for HtmlUnitDriver -->
